Moved save handlers to App.vue
